+++ UnleashedRecomp/gpu/rhi/plume_d3d12.cpp
@@ -1944,9 +1944,14 @@ namespace plume {
 
         const D3D12_TEXTURE_COPY_LOCATION copyDstLocation = toD3D12(dstLocation);
         const D3D12_TEXTURE_COPY_LOCATION copySrcLocation = toD3D12(srcLocation);
-        setSamplePositions(dstLocation.texture);
+        const RenderTexture *samplePositionTexture = (dstLocation.texture != nullptr) ? dstLocation.texture : srcLocation.texture;
+        if (samplePositionTexture != nullptr) {
+            setSamplePositions(samplePositionTexture);
+        }
         d3d->CopyTextureRegion(&copyDstLocation, dstX, dstY, dstZ, &copySrcLocation, (srcBox != nullptr) ? &copyBox : nullptr);
-        resetSamplePositions();
+        if (samplePositionTexture != nullptr) {
+            resetSamplePositions();
+        }
     }
 
     void D3D12CommandList::copyBuffer(const RenderBuffer *dstBuffer, const RenderBuffer *srcBuffer) {